// launchdarkly/types/paths/api/v2/segments/{projectKey}/{environmentKey}/{segmentKey}.types.ts
export type getSegment = (
  $: OmitValueWhenNever<{
    query: never;
    querystring: never;
    path: getSegment_Path;
    headers: getSegment_Headers;
    cookie: never;
    body: never;
    context: Context;
    response: ResponseBuilderFactory<{
      200: {
        headers: never;
        requiredHeaders: never;
        content: {
          "application/json": {
            schema: UserSegment;
          };
        };
        examples: {};
      };
      401: {
        headers: never;
        requiredHeaders: never;
        content: {
          "application/json": {
            schema: UnauthorizedErrorRep;
          };
        };
        examples: {};
      };
      404: {
        headers: never;
        requiredHeaders: never;
        content: {
          "application/json": {
            schema: NotFoundErrorRep;
          };
        };
        examples: {};
      };
      429: {
        headers: never;
        requiredHeaders: never;
        content: {
          "application/json": {
            schema: RateLimitedErrorRep;
          };
        };
        examples: {};
      };
    }>;
    x: WideOperationArgument;
    proxy: (url: string) => COUNTERFACT_RESPONSE;
    auth: { apiKey: string };
    user: never;
    delay: (milliseconds: number, maxMilliseconds?: number) => Promise<void>;
    version: never;
  }>,
) => MaybePromise<COUNTERFACT_RESPONSE>;
// ...
export type patchSegment = (
  $: OmitValueWhenNever<{
    query: never;
    querystring: never;
    path: patchSegment_Path;
    headers: patchSegment_Headers;
    cookie: never;
    body: PatchWithComment;
    context: Context;
    response: ResponseBuilderFactory<{
      200: {
        headers: never;
        requiredHeaders: never;
        content: {
          "application/json": {
            schema: UserSegment;
          };
        };
        examples: {};
      };
      400: {
        headers: never;
        requiredHeaders: never;
        content: {
          "application/json": {
            schema: InvalidRequestErrorRep;
          };
        };
        examples: {};
      };
      401: {
        headers: never;
        requiredHeaders: never;
        content: {
          "application/json": {
            schema: UnauthorizedErrorRep;
          };
        };
        examples: {};
      };
      403: {
        headers: never;
        requiredHeaders: never;
        content: {
          "application/json": {
            schema: ForbiddenErrorRep;
          };
        };
        examples: {};
      };
      404: {
        headers: never;
        requiredHeaders: never;
        content: {
          "application/json": {
            schema: NotFoundErrorRep;
          };
        };
        examples: {};
      };
      409: {
        headers: never;
        requiredHeaders: never;
        content: {
          "application/json": {
            schema: StatusConflictErrorRep;
          };
        };
        examples: {};
      };
      429: {
        headers: never;
        requiredHeaders: never;
        content: {
          "application/json": {
            schema: RateLimitedErrorRep;
          };
        };
        examples: {};
      };
    }>;
    x: WideOperationArgument;
    proxy: (url: string) => COUNTERFACT_RESPONSE;
    auth: { apiKey: string };
    user: never;
    delay: (milliseconds: number, maxMilliseconds?: number) => Promise<void>;
    version: never;
  }>,
) => MaybePromise<COUNTERFACT_RESPONSE>;

/**
 * Delete a segment.
 */
export type deleteSegment = (
  $: OmitValueWhenNever<{
    query: never;
    querystring: never;
    path: deleteSegment_Path;
    headers: deleteSegment_Headers;
    cookie: never;
    body: never;
    context: Context;
    response: ResponseBuilderFactory<{
      204: {
        headers: never;
        requiredHeaders: never;
        content: never;
        examples: {};
      };
      401: {
        headers: never;
        requiredHeaders: never;
        content: {
          "application/json": {
            schema: UnauthorizedErrorRep;
          };
        };
        examples: {};
      };
      403: {
        headers: never;
        requiredHeaders: never;
        content: {
          "application/json": {
            schema: ForbiddenErrorRep;
          };
        };
        examples: {};
      };
      404: {
        headers: never;
        requiredHeaders: never;
        content: {
          "application/json": {
            schema: NotFoundErrorRep;
          };
        };
        examples: {};
      };
      409: {
        headers: never;
        requiredHeaders: never;
        content: {
          "application/json": {
            schema: StatusConflictErrorRep;
          };
        };
        examples: {};
      };
      429: {
        headers: never;
        requiredHeaders: never;
        content: {
          "application/json": {
            schema: RateLimitedErrorRep;
          };
        };
        examples: {};
      };
    }>;
    x: WideOperationArgument;
    proxy: (url: string) => COUNTERFACT_RESPONSE;
    auth: { apiKey: string };
    user: never;
    delay: (milliseconds: number, maxMilliseconds?: number) => Promise<void>;
    version: never;
  }>,
) => MaybePromise<COUNTERFACT_RESPONSE>;

export type getSegment_Path = {
  /**
   * The project key
   */
  projectKey: string;
  /**
   * The environment key
   */
  environmentKey: string;
  /**
   * The segment key
   */
  segmentKey: string;
};

export type getSegment_Headers = { Authorization: string };

export type patchSegment_Path = {
  /**
   * The project key
   */
  projectKey: string;
  /**
   * The environment key
   */
  environmentKey: string;
  /**
   * The segment key
   */
  segmentKey: string;
};

export type patchSegment_Headers = { Authorization: string };

export type deleteSegment_Path = {
  /**
   * The project key
   */
  projectKey: string;
  /**
   * The environment key
   */
  environmentKey: string;
  /**
   * The segment key
   */
  segmentKey: string;
};

export type deleteSegment_Headers = { Authorization: string };